Reorganize repo + create taks with `yamls` + addind methods`generate()` & `get_json()` & `get_txt()` to generatods.
lint

Update dataset loading example and modify task table; add TODO for header/footer completion

typos + lint
