+++ README.md
@@ -1,68 +1,35 @@
-Title: TODO
+# Title: TODO
 
-Abstract: TODO
+# Abstract: TODO
 
+# Topics: TODO
 
-Topics: TODO
+# Babisteps Dataset Generator
 
+`main.py`, generates datasets of stories based on the task configurations found in the tasks folder. It performs the following operations:
 
-## SimpleTracking
-Descriptions: TODO
+- Reads task configurations from a `commons.yaml` file inside the specified tasks directory.
+- Loads task modules dynamically using functions defined in the tasks.
+- Generates stories using the tasks' generators.
+- Saves the generated stories in JSONL and TXT formats.
+- Creates a final dataset from the generated outputs.
 
-### ActosInLocation
-```bash
-python3 main.py \
-    --task simpletracking \
-    --locations bedroom,bathroom,kitchen \
-    --actors Alice,Bob,Charlie \
-    --q_stories 2 \
-    --states_lenght 3 \
-    --question where \
-    --answer unknown \
-    --verbosity info
-```
+## Command-line Arguments
 
+The following command-line arguments can be used to customize the behavior of the script:
 
-### ActorWithObject
-```bash
-python3 main.py \
-    --task simpletracking \
-    --actors Alice,Bob,Charlie \
-    --objects book,ball,pen \
-    --q_stories 2 \
-    --states_lenght 3 \
-    --question what \
-    --answer none \
-    --verbosity debug
-```
+- **`--task_path:`** Root path where the task configuration folders are located. 
+- **`--tasks:`** List of task names to run the experiments. Use a comma-separated string (e.g., `--tasks task1,task2`).
+- **`--output_path:`** Path to save the script results. The output folder will be timestamped.
+- **`--seed:`** Random seed for Python's random module.
+- **`--numpy_random_seed:`** Random seed for NumPy's random module.
+- **`--verbosity:`** or **`-v`** Controls the reported logging error level. (`CRITICAL|ERROR|WARNING|INFO|DEBUG`)
 
+## Example Usage
 
-## ComplexTracking
+Below are a few examples on how you might run the script:
 
-### ObjectInLocationPolar
-
-```bash
-python3 main.py \
-    --task complextracking \
-    --locations bedroom,bathroom,kitchen \
-    --actors Alice,Bob,Charlie \
-    --objects book,ball,pen \
-    --q_stories 3 \
-    --states_lenght 20 \
-    --question polar \
-    --answer yes
-```
-
-### ObjectInLocationWhat
-
-```bash
-python3 main.py \
-    --task complextracking \
-    --locations x_0,x_1,x_2,x_3 \
-    --actors y_0,y_1,y_2,y_3,y_4,y_5 \
-    --objects z_0,z_1,z_2,z_3,z_4,z_5,z_6,z_7,z_8,z_9,z_10 \
-    --q_stories 15 \
-    --states_lenght 20 \
-    --question where \
-    --answer designated_location
+### Run All Tasks with Default Settings
+```sh
+python3 main.py
 ```
