+++ babisteps/tasks/simpletracking/README.md
@@ -1,0 +1,5 @@
+* `task_name`: name of the taks
+* `entities`: # of objects in the story
+* `coordenates`: # of actors in the story
+* `gen_kwargs`: Set of possible generation probabilities
+* `func`: name of the function returning an iterable of story generators (like utils.<itarable of generators>)
