+++ babisteps/tasks/simpletracking/config.yaml
@@ -1,0 +1,5 @@
+task_name: simpletracking
+entities: 6
+coordenates: 3
+gen_kwargs:
+func: utils._get_generators
