+++ main.py
@@ -7,9 +7,12 @@ from datetime import datetime
 from pathlib import Path
 
 import numpy as np
+import yaml
 
+from babisteps import datasets as ds
 from babisteps import logger
 from babisteps import proccesing as proc
+from babisteps import utils as ut
 
 
 def split_or_empty(value):
@@ -18,50 +21,36 @@ def split_or_empty(value):
 
 def main():
     parser = argparse.ArgumentParser(description="Create a dataset of stories")
-    parser.add_argument("--q_stories",
-                        type=int,
-                        help="Number of stories to create")
-    parser.add_argument("--task",
-                        choices=["simpletracking", "complextracking"])
-    parser.add_argument("--states_lenght",
-                        type=int,
-                        help="Lenght of states by story")
-    parser.add_argument("--locations",
-                        help="List of locations",
-                        type=split_or_empty,
-                        default=[])
-    parser.add_argument("--actors",
-                        help="List of actors",
-                        type=split_or_empty,
-                        default=[])
-    parser.add_argument("--objects",
-                        help="List of objects",
-                        type=split_or_empty,
-                        default=[])
-    parser.add_argument("--question", type=str, help="Question to ask")
-    parser.add_argument("--answer", type=str, help="Answer to the question")
-    parser.add_argument("--path",
-                        type=str,
-                        default="./outputs",
-                        help="Path to save the script results")
-    # From https://github.com/EleutherAI/lm-evaluation-harness/blob/main/lm_eval/__main__.py
     parser.add_argument(
-        "--gen_kwargs",
+        "--task_path",
         type=str,
-        default=None,
-        help=("String arguments for story generation,"
-              " e.g. `p_antilocation=0.5,p_object_in_actor=0.75`"),
+        default="./babisteps/tasks",
+        help="Root path where are placed the tasks",
+    )
+    parser.add_argument(
+        "--tasks",
+        type=split_or_empty,
+        default=["all"],
+        help="List of the taks name to run the experiments",
+    )
+    parser.add_argument(
+        "--output_path",
+        type=str,
+        default="./outputs",
+        help="Path to save the script results",
     )
     parser.add_argument(
         "--seed",
         type=int,
         default=0,
-        help="Random seed for python's random module. Default set to 0.")
+        help="Random seed for python's random module. Default set to 0.",
+    )
     parser.add_argument(
         "--numpy_random_seed",
         type=int,
         default=1234,
-        help="Random seed for numpy's random module. Default set to 1234.")
+        help="Random seed for numpy's random module. Default set to 1234.",
+    )
     parser.add_argument(
         "--verbosity",
         "-v",
@@ -73,7 +62,7 @@ def main():
     args = parser.parse_args()
 
     # Get base output folder path
-    output_path = Path(args.path)
+    output_path = Path(args.output_path)
     if not os.path.exists(output_path):
         os.mkdir(output_path)
     # Add date string
@@ -82,13 +71,39 @@ def main():
         os.path.join(output_path, current_date.strftime("%Y-%m-%dT%H:%M:%S")))
     if not os.path.exists(output_path):
         os.mkdir(output_path)
-    # init of excecution
+    dataset_path = Path(os.path.join(output_path, "dataset"))
+    if not os.path.exists(dataset_path):
+        os.mkdir(dataset_path)
+    jsonl_path = Path(os.path.join(output_path, "jsonl"))
+    if not os.path.exists(jsonl_path):
+        os.mkdir(jsonl_path)
+    # init of jsonl_path
     start_time = time.time()
-    main_logger = logger.get_logger("main",
-                                    level=getattr(logging,
-                                                  f"{args.verbosity}"),
-                                    log_file=os.path.join(
-                                        output_path, "logs.txt"))
+    main_logger = logger.get_logger(
+        "main",
+        level=getattr(logging, f"{args.verbosity}"),
+        log_file=os.path.join(output_path, "logs.txt"),
+    )
+    task_path = Path(args.task_path)
+    if not os.path.exists(task_path):
+        raise ValueError(f"Path {task_path} does not exist")
+    # get task_folders as the folder inside the task_path (each element as Path)
+    task_folders = [f for f in task_path.iterdir() if f.is_dir()]
+    common_yaml = task_path / "commons.yaml"
+    if not os.path.exists(common_yaml):
+        raise ValueError(f"File {common_yaml} does not exist")
+    # load common.yaml
+    with open(common_yaml) as stream:
+        try:
+            yaml_cfg = yaml.safe_load(stream)
+        except yaml.YAMLError as exc:
+            print(exc)
+            raise ValueError(f"Error loading {common_yaml}") from exc
+
+    # update yaml_cfg with output_path
+    yaml_cfg["output_path"] = jsonl_path
+    task_path_dict = ut.create_task_path_dict(args.tasks, task_folders,
+                                              main_logger)
 
     seed_message = []
     if args.seed is not None:
@@ -102,59 +117,31 @@ def main():
     if seed_message:
         main_logger.info(" | ".join(seed_message))
 
-    if args.task == "simpletracking":
-        # Count how many lists are non-empty
-        non_empty_count = sum(
-            bool(lst) for lst in [args.locations, args.actors, args.objects])
-        assert non_empty_count == 2, (
-            "The number non-empty lists from locations, actors, and objects must be 2"
-        )
-        jsonl_dataset, txt_dataset, folder_path = proc.create_simpletracking(
-            q_stories=args.q_stories,
-            states_qty=args.states_lenght,
-            locations=args.locations,
-            actors=args.actors,
-            objects=args.objects,
-            question=args.question,
-            answer=args.answer,
-            path=output_path,
-            verbosity=getattr(logging, f"{args.verbosity}"),
-            logger=main_logger,
-            gen_kwargs=args.gen_kwargs,
-        )
-    elif args.task == "complextracking":
-        # Count how many lists are non-empty
-        non_empty_count = sum(
-            bool(lst) for lst in [args.locations, args.actors, args.objects])
-        assert non_empty_count == 3, (
-            "The number non-empty lists from locations, actors, and objects must be 3"
-        )
+    jsonl_path_dict = {}
+    for task_name_i, task_path_i in task_path_dict.items():
+        main_logger.info("Running task %s", task_name_i)
+        func_i = ut.load_function_from_config(task_path_i)
+        jsonl_dataset = []
+        txt_dataset = ""
+        generators, folder_path = func_i(**yaml_cfg)
+        main_logger.info("Generatring task:", task=task_name_i)
+        for idx, g in enumerate(generators):
+            json_i, txt_i = proc._run_generation(g, yaml_cfg)
+            json_i['idx'] = idx
+            jsonl_dataset.append(json_i)
+            txt_dataset += txt_i
+        try:
+            jsonl_file_path_i = proc.save_as_jsonl(jsonl_dataset, folder_path,
+                                                   None)
+            jsonl_path_dict[task_name_i] = jsonl_file_path_i
+            proc.save_as_txt(txt_dataset, folder_path, None)
+        except Exception as e:
+            raise Exception("Error saving dataset") from e
+        main_logger.info("SUCCESS Generating task:", task=task_name_i)
 
-        jsonl_dataset, txt_dataset, folder_path = proc.create_complextracking(
-            q_stories=args.q_stories,
-            states_qty=args.states_lenght,
-            locations=args.locations,
-            actors=args.actors,
-            objects=args.objects,
-            question=args.question,
-            answer=args.answer,
-            path=output_path,
-            verbosity=getattr(logging, f"{args.verbosity}"),
-            logger=main_logger,
-            gen_kwargs=args.gen_kwargs,
-        )
-
-    else:
-        raise ValueError("At least two lists must be non-empty")
-
-    # At this point the jsonl_dataset and folder_path are defined
-    # Time to save the dataset
-
-    try:
-        proc.save_as_jsonl(jsonl_dataset, folder_path, main_logger)
-        proc.save_as_txt(txt_dataset, folder_path, main_logger)
-    except Exception as e:
-        main_logger.exception("Error", exception=e)
+    main_logger.info("STARTING DATASET CREATION")
+    ds.create_babisteps_dataset(dataset_path, jsonl_path_dict, main_logger)
+    main_logger.info("SUCCESS DATASET CREATION")
 
     end_time = time.time()
     main_logger.info("Execution time", time=end_time - start_time)
